backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
gunicorn
SQLAlchemy
psycopg2-binary
